Add points module
