+++ runtime/src/bank.rs
@@ -156,8 +156,9 @@ use {
         },
         transaction_context::{TransactionAccount, TransactionReturnData},
     },
-    solana_stake_program::stake_state::{
-        self, InflationPointCalculationEvent, PointValue, StakeStateV2,
+    solana_stake_program::{
+        points::{InflationPointCalculationEvent, PointValue},
+        stake_state::{self, StakeStateV2},
     },
     solana_svm::{
         account_loader::{TransactionCheckResult, TransactionLoadResult},
@@ -2989,7 +2990,7 @@ impl Bank {
                         return 0;
                     };
 
-                    stake_state::calculate_points(
+                    solana_stake_program::points::calculate_points(
                         stake_account.stake_state(),
                         vote_state,
                         stake_history,
@@ -3026,7 +3027,7 @@ impl Bank {
                     delegations
                         .par_iter()
                         .map(|(_stake_pubkey, stake_account)| {
-                            stake_state::calculate_points(
+                            solana_stake_program::points::calculate_points(
                                 stake_account.stake_state(),
                                 vote_state,
                                 stake_history,
